xmpp: Fix the Stream Management implementation
There was a bug that on resumption messages would be received again.
The cause of this was probably that we send a <resume /> with an
h equal to the stanzas we sent. If we sent and received an equal amount,
no problem. But if we receive more than we sent, then duplicates will
start flooding in.