service: Fix the app getting stuck in release mode
It turns out that a release build is much faster than a
development build. The issue was that the UI sent the
[PerformPreStartCommand] too early as the background service
still had work to do before being considered ready.

As a solution we now either wait for the event to arrive by
itself, as the service sends it at the end of its initilization, or
after an explicit request if the service is already running.