fix(android): (Hopefully) fix build issues
Okay, so it turns out we cannot just have a plugin right
next to the actual app because then Flutter won't generate
both the GeneratedPluginRegistrant and the correct links to
the plugins. As such, we have to move the plugin code out of
Moxxy into the new moxxy_native library.

However, we need access to the activity to receive the notification
tap events (at least on Android). To fix this, we have a quirk that
checks if we have a notification event before finishing the handling
of the PreStartDoneEvent.