fix(ui,service): Fix massive storage leak
It turns out that picked files are not cleaned up and just accumulate.
This commit removes then if the file sending process is cancelled or if
the file is uploaded. And as a safety measure, this currenly only
happens on Android, where we know we copy files into the cache
directory.